requirements.txt
bson==0.5.8
dataclasses==0.7
Django==2.1.2
djongo==1.3.1
numpy==1.18.2
opencv-python==4.2.0.32
Pillow==7.0.0
pymongo==3.10.1
pytesseract==0.3.3
python-dateutil==2.8.1
pytz==2019.3
six==1.14.0
sqlparse==0.2.4